fix vfw_allocator type

+++ vfw_utils/include/vfw_allocator.h
@@ -1,21 +1,21 @@
 #ifndef __VFW_ALLOCATOR_H__
 #define __VFW_ALLOCATOR_H__
 #include <stdint.h>
-typedef struct 
+typedef struct
 {
-    uint32_t inner[8];
-}allocator;
+    uintptr_t inner[8];
+} allocator;
 
-typedef struct 
+typedef struct
 {
-    void* ptr;
-    uint32_t size;
+    void *ptr;
+    uintptr_t size;
 } al_block;
 
 extern void allocator_init(allocator *a);
-extern void allocator_add_block(allocator *a, uint32_t start, uint32_t size);
-extern void *allocator_alloc(allocator *a, uint32_t size, uint32_t align);
-extern void allocator_free(allocator *a, uint32_t start, uint32_t size);
+extern void allocator_add_block(allocator *a, uintptr_t start, uintptr_t size);
+extern void *allocator_alloc(allocator *a, uintptr_t size, uintptr_t align);
+extern void allocator_free(allocator *a, uintptr_t start, uintptr_t size);
 
 #define al_new() (           \
     {                        \
@@ -24,14 +24,14 @@ extern void allocator_free(allocator *a, uint32_t start, uint32_t size);
         a;                   \
     })
 #define al_add_block(a, st, sz) allocator_add_block(a, (st), (sz))
-#define al_alloc(a, sz, al) (                                          \
-    {                                                                  \
-        al_block block = {                                             \
-            .ptr = allocator_alloc(a, (uint32_t)(sz), (uint32_t)(al)), \
-            .size = (uint32_t)(sz)};                                   \
-        block;                                                         \
+#define al_alloc(a, sz, al) (                                            \
+    {                                                                    \
+        al_block block = {                                               \
+            .ptr = allocator_alloc(a, (uintptr_t)(sz), (uintptr_t)(al)), \
+            .size = (uintptr_t)(sz)};                                    \
+        block;                                                           \
     })
-#define al_free(a, b) allocator_free(a, (uint32_t)((b).ptr), (uint32_t)((b).size))
+#define al_free(a, b) allocator_free(a, (uintptr_t)((b).ptr), (uintptr_t)((b).size))
 
 #define al_reset(a) allocator_init(a)
